Add script to install a local version of Alpha gem
Grab the alpha release from github (the tgz is manually updated) and
extracted it into a git ignored dir (to avoid it appearing in editors)
and give instructions to add it to your Gemfile.

The alpha isn't being published on ruby gems, and the source repo isn't
a ruby gem (it's a node package that generates a ruby gem during it's
build step), so we need to point at a local copy.